refactor: Multiple extractions from SearchPage
